+++ CHANGES.md
@@ -13,6 +13,8 @@
 - **ENH: Making SAR attribute `snap_filename` public**
 - **ENH: Handling `ICEYE` pure SLC products**
 - **ENH: Allowing the user to choose if they want the GRD or SLC image for `ICEYE` products**
+- **ENH: Add the possibility to directly load the cloud cover for optical data (and add it in the band attributes)**
+- **ENH: Add the possibility to retrieve the quicklook path (if existing) and add the `plot` function allowing the user to plot the quicklook (if existing)**
 
 ### Bug Fixes